Moved IndicatorRequired as IndicatorUtilities.
