+++ ts/Stock/Indicators/TEMA/TEMAIndicator.ts
@@ -16,7 +16,7 @@ import type {
 } from './TEMAOptions';
 import type TEMAPoint from './TEMAPoint';
 
-import RequiredIndicatorMixin from '../../../Mixins/IndicatorRequired.js';
+import RequiredIndicatorMixin from '../IndicatorUtilities.js';
 import SeriesRegistry from '../../../Core/Series/SeriesRegistry.js';
 const {
     seriesTypes: {
